trunc pixelpos to int for drawstring

+++ snap-graph-builder/src/main/java/org/esa/snap/gpf/ui/worldmap/NestWorldMapPane.java
@@ -466,8 +466,8 @@ public class NestWorldMapPane extends JPanel {
             g2d.setColor(Color.black);
 
             g2d.drawString(text,
-                    textCenter.x - fontMetrics.stringWidth(text) / 2.0f,
-                    textCenter.y + fontMetrics.getAscent() / 2.0f);
+                           (int)textCenter.x - fontMetrics.stringWidth(text) / 2.0f,
+                           (int)textCenter.y + fontMetrics.getAscent() / 2.0f);
             g2d.setColor(color);
         }
